+++ src/features/meetings/weekly_schedules/midweek_meeting/useMidweekMeeting.tsx
@@ -2,9 +2,10 @@ import { useMemo, useState } from 'react';
 import { useRecoilValue } from 'recoil';
 import { useAppTranslation, useIntersectionObserver } from '@hooks/index';
 import { schedulesState } from '@states/schedules';
-import { getWeekDate } from '@utils/date';
+import { addMonths, generateDateFromTime, getWeekDate } from '@utils/date';
 import { formatDate } from '@services/dateformat';
 import {
+  hour24FormatState,
   midweekMeetingClassCountState,
   midweekMeetingOpeningPrayerAutoAssign,
   midweekMeetingTimeState,
@@ -38,6 +39,7 @@ const useMidweekMeeting = () => {
   const userUID = useRecoilValue(userLocalUIDState);
   const pgmStart = useRecoilValue(midweekMeetingTimeState);
   const lang = useRecoilValue(JWLangState);
+  const use24 = useRecoilValue(hour24FormatState);
   const openingPrayerAuto = useRecoilValue(
     midweekMeetingOpeningPrayerAutoAssign
   );
@@ -48,11 +50,17 @@ const useMidweekMeeting = () => {
     return schedules.length === 0;
   }, [schedules]);
 
+  const filteredSchedules = useMemo(() => {
+    const minDate = formatDate(addMonths(new Date(), -2), 'yyyy/MM/dd');
+
+    return schedules.filter((record) => record.weekOf >= minDate);
+  }, [schedules]);
+
   const week = useMemo(() => {
     if (typeof value === 'boolean') return null;
 
-    return schedules.at(value)?.weekOf || null;
-  }, [value, schedules]);
+    return filteredSchedules.at(value)?.weekOf || null;
+  }, [value, filteredSchedules]);
 
   const schedule = useMemo(() => {
     return schedules.find((record) => record.weekOf === week);
@@ -162,22 +170,31 @@ const useMidweekMeeting = () => {
   const partTimings = useMemo(() => {
     if (!schedule && !source) return;
 
+    let meetingStart = pgmStart;
+
+    if (!use24) {
+      const date = generateDateFromTime(pgmStart);
+      meetingStart = formatDate(date, 'h:mm');
+    }
+
     const result = schedulesMidweekGetTiming({
       schedule,
       dataView,
-      pgmStart,
+      pgmStart: meetingStart,
       source,
       lang,
     });
 
     return result;
-  }, [schedule, source, dataView, pgmStart, lang]);
+  }, [schedule, source, dataView, pgmStart, lang, use24]);
 
   const handleGoCurrent = () => {
     const now = getWeekDate();
     const weekOf = formatDate(now, 'yyyy/MM/dd');
 
-    const index = schedules.findIndex((record) => record.weekOf === weekOf);
+    const index = filteredSchedules.findIndex(
+      (record) => record.weekOf === weekOf
+    );
 
     setValue(index);
   };