+++ src/services/worker/backupUtils.ts
@@ -1238,7 +1238,9 @@ export const dbExportDataBackup = async (backupData: BackupDataType) => {
 
         for (const report of remoteReports) {
           const findReport = cong_field_service_reports.find(
-            (record) => record.report_id === report.report_id
+            (record) =>
+              record.report_data.person_uid === report.person_uid &&
+              record.report_data.report_date === report.report_month
           );
 
           if (!findReport) {
@@ -1456,26 +1458,28 @@ export const dbExportDataBackup = async (backupData: BackupDataType) => {
 
       obj.app_settings = { user_settings: userSettings };
 
-      const person = {
-        person_uid: myPerson.person_uid,
-        person_data: {
-          timeAway: myPerson.person_data.timeAway.filter(
-            (record) => !record._deleted
-          ),
-          emergency_contacts: myPerson.person_data.emergency_contacts.filter(
-            (record) => !record._deleted
-          ),
-        },
-      };
+      if (myPerson) {
+        const person = {
+          person_uid: myPerson.person_uid,
+          person_data: {
+            timeAway: myPerson.person_data.timeAway.filter(
+              (record) => !record._deleted
+            ),
+            emergency_contacts: myPerson.person_data.emergency_contacts.filter(
+              (record) => !record._deleted
+            ),
+          },
+        };
 
-      encryptObject({
-        data: person,
-        table: 'persons',
-        masterKey,
-        accessCode,
-      });
+        encryptObject({
+          data: person,
+          table: 'persons',
+          masterKey,
+          accessCode,
+        });
 
-      obj.persons = [person];
+        obj.persons = [person];
+      }
     }
 
     // include publisher bible studies and field reports