+++ CHANGELOG.md
@@ -1,3 +1,70 @@
+# [3.3.0](https://github.com/sws2apps/organized-app/compare/v3.2.0...v3.3.0) (2024-12-08)
+
+
+### Bug Fixes
+
+* **app:** persist some app data when logging out ([e4cd3d1](https://github.com/sws2apps/organized-app/commit/e4cd3d1875d3c1d1ccc5cd50907e89e6370c16c3))
+* **components:** update margin for typography in assignments checklist ([082ca1e](https://github.com/sws2apps/organized-app/commit/082ca1e172bcb529a260e27cd887a05cf5b2d9e4))
+* **congregation:** unwanted error message for changing codes ([2bc7361](https://github.com/sws2apps/organized-app/commit/2bc73615c533f17651264ae3a9df5607a95cf3f4))
+* **dashboard:** manual backup not running ([3c9a449](https://github.com/sws2apps/organized-app/commit/3c9a449b92dc56066e8fae8e073a6fd4c36589b8))
+* **locales:** issue with tagalog language translation ([04823d9](https://github.com/sws2apps/organized-app/commit/04823d9d71a9c25bcfc91bc12896e32d1c05568c))
+* **locales:** update language path for hungarian ([8335cf6](https://github.com/sws2apps/organized-app/commit/8335cf6dd5a23c5f64a292f1efc2208c7530442b))
+* **ministry:** month selector in dark them for pioneer applications ([4d367c9](https://github.com/sws2apps/organized-app/commit/4d367c9637c305106ba510a0968f3657276f45c8))
+* **ministry:** report submission flow ([4d2a285](https://github.com/sws2apps/organized-app/commit/4d2a285302f8535427cd7a30da91093f15e6519c))
+* **persons:** update date of baptism change flow ([4a60e74](https://github.com/sws2apps/organized-app/commit/4a60e74d802eae732cab0e675d362b06cce5445a))
+* **reminders:** check if publisher in current report month ([adb25e4](https://github.com/sws2apps/organized-app/commit/adb25e4b91e32c7f3774b1ab3f7c5a57dd6737ce))
+* **reports:** add indeterminate state for S-21 export checkboxes ([569ad6e](https://github.com/sws2apps/organized-app/commit/569ad6e50eb880815d91d11c5e33ea892d8a1c00))
+* **reports:** remove duplicate reports ([bc90ace](https://github.com/sws2apps/organized-app/commit/bc90ace295c124223c8deb1239d1aa5c6dd339ba))
+* **reports:** update submission flow ([4f52dfb](https://github.com/sws2apps/organized-app/commit/4f52dfb3602f2f2e84c1a4dc9e18cd39488c737b))
+* **reports:** update year range when exporting congregation cards ([246e8c8](https://github.com/sws2apps/organized-app/commit/246e8c86123736d4dc045d01484f7e43ef2798f6))
+* **startup:** screens mixed up in some conditions ([662f04e](https://github.com/sws2apps/organized-app/commit/662f04e10f0bfd143bc9560db197e3b36964967b))
+* **templates:** use 24 hour format settings in exported pdf ([ccb02fa](https://github.com/sws2apps/organized-app/commit/ccb02fa7c87385c7b12e5dd5ca1fd3c908fbf6c3))
+
+
+### Features
+
+* **meetings:** add 'No assignments yet' in the assignment history ([41b2955](https://github.com/sws2apps/organized-app/commit/41b2955a4c69ef7eaf88a8c41d2c0c673c4fce0e))
+
+# [3.2.0](https://github.com/sws2apps/organized-app/compare/v3.1.0...v3.2.0) (2024-12-03)
+
+
+### Bug Fixes
+
+* **app:** improve buttons, fields, and export visuals ([0923a74](https://github.com/sws2apps/organized-app/commit/0923a74508c22ff808d4b92e688f0e25219088c8))
+* **app:** source material for other users ([1e858f1](https://github.com/sws2apps/organized-app/commit/1e858f16e652a99729e401c628af351ff01fc97f))
+* **components:** manual input and key down event for date picker ([a87316a](https://github.com/sws2apps/organized-app/commit/a87316a5382369b59f5a57e1e3538803c0aeda03))
+* **components:** manual input and key down event for date picker ([5870c86](https://github.com/sws2apps/organized-app/commit/5870c864c387de107c2af2da8c9297a79da9f501))
+* **components:** unfreeze top section for new update ([7e2e23b](https://github.com/sws2apps/organized-app/commit/7e2e23bf46c2009be7fa4eb0b38a60fcd9492b51))
+* **dashboard:** update users allowed to open field service groups ([7997fb7](https://github.com/sws2apps/organized-app/commit/7997fb7ef15417f22ced5e6016c9c3235ad115cc))
+* **locales:** updated translation from Crowdin ([caa4461](https://github.com/sws2apps/organized-app/commit/caa4461eacfbe50d8bf4beb497da50a5624c4f2b))
+* **meetings:** months list when publishing midweek schedules ([cc852fd](https://github.com/sws2apps/organized-app/commit/cc852fdbec0c1d3a274818a4016815b675db2933))
+* **mettings:** weeks list when viewing only midweek schedules ([5ee25bd](https://github.com/sws2apps/organized-app/commit/5ee25bd5ec0d121f4b2f218d611c73d6e1f1d94c))
+* **schedules:** include current week in active assignments ([114ed37](https://github.com/sws2apps/organized-app/commit/114ed374ac757f967af12db4b175970f627cf12e))
+* **startup:** fix infinite loading in some scenario ([2d84f18](https://github.com/sws2apps/organized-app/commit/2d84f18dae6ae774fb81d3acf141f05766cd2e35))
+* **templates:** update midweek schedule and assignment slips ([dafb2f5](https://github.com/sws2apps/organized-app/commit/dafb2f5873c9ebb71330554101fd9df443ef9414))
+* **whatsnew:** use new release notes format ([c83f65a](https://github.com/sws2apps/organized-app/commit/c83f65a735b88c461d8f499683f384d7b560775d))
+
+
+### Features
+
+* **app:** add first month of report in person details ([764e58d](https://github.com/sws2apps/organized-app/commit/764e58df027ee83bb9a6f757fbc72c89cdd9024f))
+* **locales:** add estonian language ([dfd6633](https://github.com/sws2apps/organized-app/commit/dfd6633718b1965917a3af9ab64a6f7571615085))
+
+# [3.1.0](https://github.com/sws2apps/organized-app/compare/v3.0.0...v3.1.0) (2024-12-01)
+
+### Bug Fixes
+
+- **app:** add missing color for dark theme ([9578468](https://github.com/sws2apps/organized-app/commit/9578468f10c7b2ac38f39da386306a9d33fc5716))
+- **app:** move code for setting default theme ([7916b6d](https://github.com/sws2apps/organized-app/commit/7916b6d300aa7f228316e1470d4094c63c5a2e62))
+- **features:** localize some terms in session details ([89911b7](https://github.com/sws2apps/organized-app/commit/89911b78443446d5bf91f192a7eecfe8f4b94d5b))
+- **schedules:** student selection for some part type ([e22094a](https://github.com/sws2apps/organized-app/commit/e22094afc8eb680acef5775171b860712950b742))
+- **templates:** get correct speaker name in weekend schedule ([ddd6116](https://github.com/sws2apps/organized-app/commit/ddd6116e611fff7aaacae4e0f77bc8fe285a8868))
+
+### Features
+
+- **deps:** bump @mui/x-tree-view from 7.22.1 to 7.23.0 ([7cb8894](https://github.com/sws2apps/organized-app/commit/7cb88947f1ddeb23731acb69a575b29a7d50de10))
+- **deps:** bump jw-epub-parser from 3.28.1 to 3.29.1 ([df00997](https://github.com/sws2apps/organized-app/commit/df0099735bfc38702b2c8588c2480629f3cdf563))
+
 # [3.0.0] (2024-12-01)
 
 ## The Organized App is released today!